job-search/scraper/requirements.txt
python-jobspy>=1.1.0
pyyaml>=6.0
